refac: Delete TestUtils and AndroidTestUtils
