+++ insights/src/androidTest/kotlin/com/algolia/instantsearch/AndroidTestDatabaseSharedPreferences.kt
@@ -2,10 +2,9 @@ package com.algolia.instantsearch
 
 import android.support.test.InstrumentationRegistry
 import android.support.test.runner.AndroidJUnit4
-import com.algolia.instantsearch.AndroidTestUtils.click
-import com.algolia.instantsearch.AndroidTestUtils.conversion
-import com.algolia.instantsearch.AndroidTestUtils.view
+import com.algolia.instantsearch.insights.converter.ConverterEventToEventInternal
 import com.algolia.instantsearch.insights.database.DatabaseSharedPreferences
+import com.algolia.instantsearch.insights.event.Event
 import org.junit.Test
 import org.junit.runner.RunWith
 import kotlin.test.assertTrue
@@ -15,6 +14,43 @@ import kotlin.test.assertTrue
 class AndroidTestDatabaseSharedPreferences {
 
     private val context get() = InstrumentationRegistry.getContext()
+    private val eventA = "EventA"
+    private val eventB = "EventB"
+    private val eventC = "EventC"
+    private val indexName = "latency"
+    private val queryId = "6de2f7eaa537fa93d8f8f05b927953b1"
+    private val userToken = "foobarbaz"
+    private val positions = listOf(1)
+    private val objectIDs = listOf("54675051")
+    private val timestamp = System.currentTimeMillis()
+    private val eventClick = Event.Click(
+        eventA,
+        indexName,
+        userToken,
+        timestamp,
+        queryId,
+        objectIDs,
+        positions
+    )
+    private val eventConversion = Event.Conversion(
+        eventB,
+        indexName,
+        userToken,
+        timestamp,
+        queryId,
+        objectIDs
+    )
+    private val eventView = Event.View(
+        eventC,
+        indexName,
+        userToken,
+        timestamp,
+        queryId,
+        objectIDs
+    )
+    private val click = ConverterEventToEventInternal.convert(eventClick)
+    private val view = ConverterEventToEventInternal.convert(eventView)
+    private val conversion = ConverterEventToEventInternal.convert(eventConversion)
 
     @Test
     fun test() {
@@ -22,7 +58,7 @@ class AndroidTestDatabaseSharedPreferences {
             click,
             conversion
         )
-        val database = DatabaseSharedPreferences(context, AndroidTestUtils.indexName)
+        val database = DatabaseSharedPreferences(context, indexName)
 
         database.overwrite(events)
